SC-11043 moved serializer code to serializer package
