+++ plugins/outputs/sematext/serializer/serializer_test.go
@@ -1,4 +1,4 @@
-package sematext
+package serializer
 
 import (
 	"fmt"
